harvestdate for URL moved to frontend services
